refactor: remove unnecessary ovh prefix
