+++ .babelrc
@@ -1,5 +1,5 @@
 {
-    "plugins": [
-        "@babel/plugin-syntax-dynamic-import"
-    ]
+  "plugins": [
+    "@babel/plugin-syntax-dynamic-import"
+  ]
 }
