+++ package.json
@@ -1,22 +1,22 @@
 {
-  "name": "ovh-manager",
+  "name": "manager",
   "version": "0.0.0",
+  "private": true,
   "description": "OVH Control Panel also known as Manager",
   "license": "BSD-3-Clause",
-  "private": true,
-  "engines": {
-    "node": ">=8.11"
-  },
+  "workspaces": [
+    "packages/*"
+  ],
   "scripts": {
-    "build": "webpack --config ./packages/ovh-manager/webpack.config.js --env=production",
-    "start": "webpack-dev-server --config ./packages/ovh-manager/webpack.config.js",
-    "test": "yarn lint",
+    "build": "webpack --config ./packages/manager-layout-ovh/webpack.config.js --env=production",
     "links": "ls -l node_modules/ | grep ^l || echo 'no linked packages'",
     "lint": "run-p lint:css lint:html lint:js lint:md",
     "lint:css": "stylelint 'packages/**/*.less' --fix",
     "lint:html": "htmlhint 'packages/**/*.html'",
     "lint:js": "eslint --fix .",
     "lint:md": "remark -qf .",
+    "start": "webpack-dev-server --config ./packages/manager-layout-ovh/webpack.config.js",
+    "test": "yarn lint",
     "version": "conventional-changelog -p angular -i CHANGELOG.md -s && git add CHANGELOG.md"
   },
   "devDependencies": {
@@ -34,12 +34,7 @@
     "stylelint": "^9.5.0",
     "stylelint-config-standard": "^18.2.0"
   },
-  "workspaces": [
-    "packages/*"
-  ],
-  "resolve": {
-    "alias": {
-      "jquery": "jquery/src/jquery"
-    }
+  "engines": {
+    "node": ">=8.11"
   }
 }
