+++ packages/manager-core/src/manager-core.constants.js
