+++ packages/manager-core/src/session/session.service.js
