+++ packages/manager-layout-ovh/src/index.js
@@ -7,10 +7,10 @@ import set from 'lodash/set';
 import ssoAuth from 'ovh-angular-sso-auth';
 import OvhHttp from 'ovh-angular-http';
 
-import core from '@ovh-ux/ovh-manager-core'; // eslint-disable-line import/no-extraneous-dependencies
-import welcome from '@ovh-ux/ovh-manager-welcome'; // eslint-disable-line import/no-extraneous-dependencies
+import core from '@ovh-ux/manager-core'; // eslint-disable-line import/no-extraneous-dependencies
+import welcome from '@ovh-ux/manager-welcome'; // eslint-disable-line import/no-extraneous-dependencies
 
-import routing from './ovh-manager.routes';
+import routing from './manager-layout-ovh.routes';
 import sidebar from './sidebar';
 import navbar from './navbar';
 
@@ -18,8 +18,8 @@ import 'ovh-angular-otrs';
 import 'ovh-ui-angular';
 import 'bootstrap';
 
-import './ovh-manager.less';
-import './ovh-manager.scss';
+import './manager-layout-ovh.less';
+import './manager-layout-ovh.scss';
 
 angular
   .module('ovhManager', [
