+++ packages/manager-layout-ovh/src/manager-layout-ovh.scss
@@ -1,7 +1,7 @@
-@import "../../node_modules/bootstrap4/scss/_functions";
-@import "../../node_modules/bootstrap4/scss/_variables";
-@import "../../node_modules/bootstrap4/scss/_mixins";
-@import '../../node_modules/bootstrap4/scss/_utilities.scss';
+@import "../../../node_modules/bootstrap4/scss/_functions";
+@import "../../../node_modules/bootstrap4/scss/_variables";
+@import "../../../node_modules/bootstrap4/scss/_mixins";
+@import '../../../node_modules/bootstrap4/scss/_utilities.scss';
 
 // Microsoft Icons
 // @import './../../../node_modules/office-ui-fabric-react/dist/sass/variables/_Icon.Variables.scss';
