+++ packages/manager-layout-ovh/src/navbar/navbar-notification.service.js
