+++ packages/manager-layout-ovh/src/navbar/navbar.controller.js
