+++ packages/manager-layout-ovh/src/navbar/navbar.less
