+++ packages/manager-layout-ovh/src/navbar/navbar.service.js
