+++ packages/manager-layout-ovh/src/navbar/translations/Messages_fr_FR.xml
