+++ packages/manager-layout-ovh/src/sidebar/sidebar.constants.js
