+++ packages/manager-layout-ovh/src/sidebar/sidebar.controller.js
