+++ packages/manager-layout-ovh/src/sidebar/sidebar.html
