+++ packages/manager-layout-ovh/src/sidebar/translations/Messages_fr_FR.xml
