+++ packages/manager-layout-ovh/webpack.config.js
