+++ packages/manager-layout-ovh/webpack.prod.js
