+++ packages/manager-welcome/README.md
@@ -1,0 +1,1 @@
+# manager-welcome
