+++ packages/manager-welcome/src/index.js
@@ -1,9 +1,9 @@
 import angular from 'angular';
 import uiRouter from '@uirouter/angularjs';
 import ocLazyLoad from 'oclazyload';
-import core from '@ovh-ux/ovh-manager-core'; // eslint-disable-line import/no-extraneous-dependencies
+import core from '@ovh-ux/manager-core'; // eslint-disable-line import/no-extraneous-dependencies
 
-import routing from './ovh-manager-welcome.routes';
+import routing from './manager-welcome.routes';
 
 const moduleName = 'ovhManagerWelcome';
