+++ packages/manager-welcome/src/manager-welcome.component.js
@@ -1,5 +1,5 @@
 import angular from 'angular';
-import template from './ovh-manager-welcome.html';
+import template from './manager-welcome.html';
 
 export default angular
   .module('OvhManagerWelcomeComponent', [])
